moving error handler to server file
