--- src/constants/error.js
@@ -1,16 +1,0 @@
-// eslint-disable-next-line complexity
-export function apiErrorAlert(error, defaultMsg = 'There was a problem.') {
-  console.log(defaultMsg, error); // eslint-disable-line no-console
-
-  let message = defaultMsg
-
-  if (
-    error.response &&
-    error.response.data &&
-    error.response.data.error
-  ) {
-    message = error.response.data.error;
-  }
-
-  alert(message)
-}
